+++ docs/conf.py
@@ -26,10 +26,10 @@ if os.getenv('SPELLCHECK'):
 source_suffix = '.rst'
 master_doc = 'index'
 project = 'explorepy'
-year = '2018-2020'
+year = '2018-2022'
 author = 'Mentalab GmbH.'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '1.6.3'
+version = release = '1.7.0'
 
 pygments_style = 'trac'
 templates_path = ['.']